Refactor style and make code more modular for separation of concern
