+++ src/components/DefaultLayout.tsx
@@ -19,7 +19,7 @@ const DefaultLayout = () => {
           <div className="flex h-screen overflow-hidden">
             <div className="relative flex flex-1 flex-col overflow-y-auto overflow-x-hidden">
               <main>
-                <div className="mx-auto max-w-screen-2xl px-4 md:px-6 2xl:px-4">
+                <div className="mx-auto max-w-screen-2xl">
                   <Suspense fallback={<Spinner />}>
                     <Outlet />
                   </Suspense>
